Minor fix in ComicModel.kt model

+++ data_network/src/main/java/com/thanosfisherman/data_network/models/response/comics/Result.kt
@@ -16,7 +16,7 @@ data class Result(
     val images: List<Image>,
     val isbn: String,
     val issn: String,
-    val issueNumber: Int,
+    val issueNumber: Double,
     val modified: String,
     val pageCount: Int,
     val prices: List<Price>,
