+++ domain/src/main/java/com/thanosfisherman/domain/model/ComicModel.kt
@@ -3,7 +3,7 @@ package com.thanosfisherman.domain.model
 data class ComicModel(
     val id: Long,
     val title: String,
-    val issueNumber: Int,
+    val issueNumber: Double,
     val description: String,
     val pic: String
 )
